pyproject.toml
[project]
name = "manage-skills"
version = "0.1.0"
description = "SRE Skills Manager"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.1",
    "pydantic>=2.7.0",
    "rich>=13.7.1"
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
strict = true
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
